requirements.txt
asgiref==3.10.0
Django==5.2.7
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
